+++ CHANGELOG.md
@@ -10,12 +10,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - The compiler will now warn if additional text is present on the same line before or after commands and other statements.
   - For example, the following code will emit a warning, because the last `>` character is likely a typo: `<<wait 5>>>`
+- Added a new method, `Dialogue.GetHeaders`, which returns the collection of headers present on a node.
+- Added a new method, `Dialogue.GetHeaderValue`, which returns the value of the specified header on a node.
 
 ### Changed
 
 - Commands are now better at checking to see if the first word is a keyword (e.g. `return`) or a word that just _begins_ with a keyword (`returnToMenu`).
 
 ### Removed
+
+- Removed `GetTagsForNode`. This method is replaced with `GetHeaderValue(nodeName, "tags")`.
 
 ## [3.0.0-beta1] 2024-11-29